+++ pyproject.toml
@@ -24,6 +24,7 @@ torch = "2.7.1"
 transformers = "4.52.4"
 jupyter = "^1.1.1"
 ipykernel = "^6.29.5"
+gradio = "^5.33.2"
 
 
 [build-system]
